pages/requirements.txt
streamlit
pandas
matplotlib
seaborn
PIL
numpygit
sklearn
